fix(channel-gateway): extract config redaction under 900-line budget
Move redaction helpers to config-redaction.ts so config.ts stays within
the modularity source budget while keeping public re-exports stable.
